Refactor NN: float arrays & parallel training support
Converted all NN data structures from double[] to float[] for improved performance and memory efficiency.
Introduced parallelized versions of FeedForward and Train methods (FeedForward_Parallel, Train_Parallel) using Parallel.For.
Updated all training and inference code to use the new float-based and parallel methods.
Adjusted MNIST and XOR data handling, normalization, and targets to float.
Added new XOR3Learn project and fixed project naming in the solution.
Overall, enables faster, more efficient training and inference leveraging multi-core CPUs.

+++ CSharpBible/Graphics/XOR3Learn/Program.cs
@@ -30,7 +30,7 @@ class Program
         NeuralNetwork nn = IoC.GetRequiredService<NeuralNetwork>();
 
         // 2. Trainingsdaten (XOR-Logik)
-        double[][] inputs =
+        float[][] inputs =
         [
             [0, 0,0],
             [0, 1 ,0],
@@ -42,7 +42,7 @@ class Program
             [1, 1,1]
         ];
 
-        double[][] targets =
+        float[][] targets =
         [
             [0],
             [1],
@@ -71,7 +71,7 @@ class Program
                 var result = 1.0;
                 foreach (var tuple in inputs.Zip(targets))
                 {
-                    double[] output = nn.FeedForward(tuple.First);
+                    float[] output = nn.FeedForward(tuple.First);
                     result *= (output[0] - 0.5) * (tuple.Second[0] - 0.5) * 4;
                 }
                 Console.WriteLine($"Zwischenergebnis: {result:F4}");
@@ -84,7 +84,7 @@ class Program
             var finalresult = 1.0;
             foreach (var tuple in inputs.Zip(targets))
             {
-                double[] output = nn.FeedForward(tuple.First);
+                float[] output = nn.FeedForward(tuple.First);
                 Console.WriteLine($"Input: {tuple.First[0]}, {tuple.First[1]}, {tuple.First[2]} | Vorhersage: {output[0]:F4}");
                 finalresult *= (output[0] - 0.5) * (tuple.Second[0] - 0.5) * 4;
             }
